reviseFollowerList, FollowerItem
