+++ src/components/FollowerList.jsx
@@ -1,8 +1,7 @@
 import styled from "styled-components"
-import { useState } from "react"
 import { FollowerItem } from '.'
 // 測試用假資料串接
-import { dummyUsers } from "testData/dummyUserInfo"
+
 
 
 const StyledFollowList = styled.div`
@@ -11,23 +10,23 @@ const StyledFollowList = styled.div`
 `
 
 
-const FollowerList = () => {
+const FollowerList = ({users, onToggleFollow}) => {
 
-    const [users, setUsers] = useState(dummyUsers)
     
-    function handleFollow(id) {
-        setUsers((users) => {
-            return users.map((user) => {
-              if(user.id === id) {
-                return {
-                 ...user,
-                 isFollow: !user.isFollow
-                }
-              }
-              return user
-            })
-        }) 
-    }
+    
+    // function handleFollow(id) {
+    //     setUsers((users) => {
+    //         return users.map((user) => {
+    //           if(user.id === id) {
+    //             return {
+    //              ...user,
+    //              isFollow: !user.isFollow
+    //             }
+    //           }
+    //           return user
+    //         })
+    //     }) 
+    // }
 
     return (
         <StyledFollowList>
@@ -36,7 +35,7 @@ const FollowerList = () => {
                    <FollowerItem 
                     key = {user.id}
                     item = {user}
-                    onToggleFollow = {handleFollow}
+                    onToggleFollow = {onToggleFollow}
                    />
                 )
             })}
