+++ src/components/UserPageFollowerList.jsx
@@ -1,5 +1,9 @@
 import { useParams } from "react-router-dom"
 import { UserHeader, TabList, TabItem, FollowerList } from "components"
+import { dummyUsers } from "testData/dummyUserInfo"
+import { useState } from "react"
+
+
 
 /**
  * [前台] 使用者資料頁（追隨者）
@@ -7,7 +11,22 @@ import { UserHeader, TabList, TabItem, FollowerList } from "components"
  */
 const UserPageFollowerList = () => {
     const { user_id } = useParams();
-
+    const [users, setUsers] = useState(dummyUsers)
+       
+    function handleFollow(id) {
+        setUsers((users) => {
+            return users.map((user) => {
+              if(user.id === id) {
+                return {
+                 ...user,
+                 isFollow: !user.isFollow
+                }
+              }
+              return user
+            })
+        }) 
+    }
+  
     return (
         <>
             <UserHeader
@@ -18,7 +37,7 @@ const UserPageFollowerList = () => {
                 <TabItem to={`/user/${user_id}/follower`} text="追隨者" />
                 <TabItem to={`/user/${user_id}/following`} text="正在追隨" />
             </TabList>
-          <FollowerList />
+          <FollowerList users={users} onToggleFollow={handleFollow}/>
         </>
     )
 }
